sandbox prevents fs read write inside Node.js REPL.
which is what it was intended to do. the way it was set up prior was addressing all of the sub methods of fs instead of only the parent object. I just failed to notice that the module  was still importable. anywho, works now, and actually made the sandbox object a lot easier to read.